01_universal_foundations/02_cross_asset_statistical_arbitrage/requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0
xgboost>=2.0.0
statsmodels>=0.14.0
matplotlib>=3.7.0
seaborn>=0.12.0
joblib>=1.3.0
tqdm>=4.65.0
PyYAML>=6.0
numba>=0.57.0
